requirements.txt
tensorflow==1.13.1
tensorflow-gpu==1.13.1
Cython
contextlib2
pillow
lxml
jupyter
matplotlib
git+https://github.com/philferriere/cocoapi.git#subdirectory=PythonAPI